+++ pytradfri/__init__.py
@@ -1,9 +1,8 @@
 """Implement an API wrapper around Ikea Tradfri."""
 
-from .api.api_utils import retry_timeout
 from .error import (
     PyTradFriError, RequestError, ClientError, ServerError, RequestTimeout)
 from .gateway import Gateway
 
 __all__ = ['Gateway', 'PyTradFriError', 'RequestError', 'ClientError',
-           'ServerError', 'RequestTimeout', 'retry_timeout']
+           'ServerError', 'RequestTimeout']
